♻️ Refactor ui components index.ts
