+++ frontend/src/common/components/ui/index.ts
@@ -1,0 +1,5 @@
+// buttons
+export * from "./Buttons/Button";
+// inputs
+export * from "./Inputs/Input";
+export * from "./Inputs/LabeledInput";
